+++ variables-naming.tf
@@ -35,3 +35,9 @@ variable "custom_ipconfig_name" {
   type        = string
   default     = null
 }
+
+variable "custom_dcr_name" {
+  description = "Custom name for Data collection rule association"
+  type        = string
+  default     = null
+}